1.code/pyproject.toml
[project]
name = "sync-hub"
version = "1.0.0"
description = "Local dev (FastAPI) + tooling for Sync Hub (uv-based)"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
  "fastapi>=0.115",
  "uvicorn>=0.30",
  "pydantic>=2.8",
  "requests>=2.32",
  "httpx>=0.27",
  "python-dotenv>=1.0",
  "boto3>=1.34"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5",          # ruff 0.4+ 에서 format 서브커맨드 지원
  "mypy>=1.10",
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "anyio>=4.4"
]
